+++ source/de/quippy/javamod/multimedia/opl3/OPL3Mixer.java
@@ -215,8 +215,8 @@ public class OPL3Mixer extends BasicMixer
 						wideRBuffer[writePointer++]=sampler;
 						if (writePointer>=maxWideStereo) writePointer=0;
 
-						sampler+=(wideLBuffer[readPointer]>>1);
-						samplel+=(wideRBuffer[readPointer++]>>1);
+						sampler+=(int)(wideLBuffer[readPointer]>>1);
+						samplel+=(int)(wideRBuffer[readPointer++]>>1);
 						if (readPointer>=maxWideStereo) readPointer=0;
 					}
